SC-56B Added edit-case component
Fixed routing to a new component and added the right features to it. Also fixed a bug where the case number was not saved correctly to the case. A small change in default-card to show case number if it is loaded in edit-case (boolean as an input variable). Also fixed links in header to "ny anmälan" and "hämta ärende".

Co-Authored-By: henrikliu <henrikliu@users.noreply.github.com>